Multiples modificaciones, ahora los campos de hora son de tipo int, cuando se envía una correo tambíen se guarda en la base de datos, implementado metodo para poder consultar todos los correos en base a unos parámtros, aunque NO TESTEADO, solucionado comprobación de averias al realizar una reserva, y arreglada la consulta de plazas para un viaje que antes no usaba la hora que se le pasaba por parámetro

+++ appFinal/src/main/java/bosonit/formacion/appFinal/reservas/domain/Autobus.java
@@ -26,7 +26,7 @@ public class Autobus {
     private Date fechaViaje;
 
     @Column
-    private long horaSalida;
+    private int horaSalida;
 
     @Column
     @NotNull
